+++ Sources/Flare/Classes/Models/StoreProductDiscount.swift
@@ -63,6 +63,10 @@ extension StoreProductDiscount: IStoreProductDiscount {
     public var price: Decimal {
         discount.price
     }
+    
+    public var localizedPriceString: String? {
+        discount.localizedPriceString
+    }
 
     public var paymentMode: PaymentMode {
         discount.paymentMode